Changes file extension
