+++ docs/fds-components/iframe.html
@@ -462,4 +462,4 @@
             
         
             
-              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.4aeb357111fb20b7c7a0.bundle.js"></script><script src="vendors~main.4aeb357111fb20b7c7a0.bundle.js"></script><script src="main.4aeb357111fb20b7c7a0.bundle.js"></script></body></html>
+              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.3d9d3919079c8c69511a.bundle.js"></script><script src="vendors~main.3d9d3919079c8c69511a.bundle.js"></script><script src="main.3d9d3919079c8c69511a.bundle.js"></script></body></html>
